Add awsome remove confirm function - wow

+++ js/app.controller.js
@@ -69,6 +69,8 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    const isConfirmed = confirm('delete?')
+    if(!isConfirmed) return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
